Added LineDetector Class
The LineDetector Class uses the modern robotics optical distance sensor (ods) to pick out the reflectivity of a colored line. The reflectivity values are stored as constants and the the .foundColoredLine() method returns back the color of the line determined by the ods. This can be used to track the white line up to the beacon,